fix images uploaders into one middleware file and fix delete user's, properties and related properties images to be done with error handling and group them in one helper-funs file

+++ helperFunctions/deleteRelatedProperies.js
@@ -3,18 +3,15 @@ const appError = require('../utils/appError.js');
 const httpStatus = require('../utils/HTTP.status.text.js')
 const fs = require ('fs');
 const path = require('path');
+const {deletedPropertyImagesCleaner} = require('./tosImageCleaner.js');
 
 module.exports = async (req,res,next)=>{
     const propstoDelete = await Propertie.find({owner:req.currentUser.id});
-    propstoDelete.forEach(propery =>{
-        const imagesFolder = path.join(__dirname,'..');
-        propery.images.forEach(image => {
-            fs.unlink(path.join(imagesFolder,image),(err)=>{
-                const error = appError.create(`error while deleting property images ${err.message}`,500,httpStatus.FAIL);
-                return next(error);                
-            });
-            console.log('deleted successfuly');            
-        });
+    if(propstoDelete.length === 0){
+        return true;
+    }
+    await propstoDelete.forEach(prop=>{
+        deletedPropertyImagesCleaner(prop.images,next);
     })
     await Propertie.deleteMany({owner:req.currentUser.id});
     return true;
